requirements.txt
numpy  
pandas  
matplotlib  
torch  
torchvision  
streamlit  
pillow  
streamlit-drawable-canvas  